move uprobe
